# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple blog service based on golang,mariadb and okteto",
        "title": "blog service",
        "contact": {
            "name": "Elchn"
        },
        "version": "v0.0"
    },
    "host": "https://blog-service-poiyyort.cloud.okteto.net",
    "basePath": "/api/v1",
    "paths": {
...
        "/upload/file": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "upload a file and get access url back",
                "parameters": [
                    {
                        "default": 1,
                        "description": "file type",
                        "name": "type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "file",
                        "name": "file",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeeded",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/gin.H"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code defines the business error code.",
                    "type": "integer"
                },
                "errors": {
                    "$ref": "#/definitions/errors.MyError"
                },
                "message": {
                    "description": "Message contains the detail of this message.\nThis message is suitable to be exposed to external",
                    "type": "string"
                }
            }
        },
...
        "app.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code defines the business error code.",
                    "type": "integer"
                },
                "data": {
                    "description": "Reference returns the reference document which maybe useful to solve this error.",
                    "type": "any"
                },
                "message": {
                    "description": "Message contains the detail of this message.\nThis message is suitable to be exposed to external",
                    "type": "string"
                }
            }
        },
        "errors.MyError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code defines the business error code.",
                    "type": "integer"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/errors.MyError"
                    }
                },
                "message": {
                    "description": "Message contains the detail of this message.\nThis message is suitable to be exposed to external",
                    "type": "string"
                },
                "specific": {
                    "type": "string"
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": {
                "type": "any"
            }
        },
...
    }
}